Switch the dimensions array to use uint32_t instead of size_t.
size_t isn't safe, since we pack/unpack the array as a 32-bit int array, but
that is the wrong type for 64-bit. Switching to uint32_t is better, since we
only support 1 dimension today, and won't need many more than that even for
complex cases in the future.

+++ cpu_ref/rsCpuIntrinsic.cpp
@@ -54,7 +54,7 @@ void RsdCpuScriptIntrinsic::setGlobalVar(uint32_t slot, const void *data, size_t
 
 void RsdCpuScriptIntrinsic::setGlobalVarWithElemDims(uint32_t slot, const void *data,
                                                      size_t dataLength, const Element *e,
-                                                     const size_t *dims, size_t dimLength) {
+                                                     const uint32_t *dims, size_t dimLength) {
     mCtx->getContext()->setError(RS_ERROR_FATAL_DRIVER,
                                  "Unexpected RsdCpuScriptIntrinsic::setGlobalVarWithElemDims");
 }
